# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: LAZI ST-Ingestion API
  description: |
    Internal API for accessing ServiceTitan data synced to the LAZI platform.
    
    This API provides read access to master tables (synced from ServiceTitan)
    and write access via the mutation queue (changes pushed back to ServiceTitan).
    
    ## Authentication
    All endpoints require an API key passed in the `x-api-key` header.
    
    ## Rate Limiting
    - 100 requests per minute per API key
    - 1000 requests per hour per API key
    
    ## Pagination
    All list endpoints support pagination via `limit` and `offset` query parameters.
    Maximum limit is 1000 records per request.
  version: 1.0.0
  contact:
    name: LAZI Support
    email: support@lazilabs.com
servers:
  - url: http://localhost:3001/api
    description: Local development
  - url: https://st-ingestion.lazilabs.com/api
    description: Production
...
paths:
...
  /v1/customers:
    get:
      tags: [Customers]
      summary: List customers
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/search'
        - $ref: '#/components/parameters/active_only'
        - name: type
          in: query
          schema:
            type: string
            enum: [Residential, Commercial]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'
  /v1/customers/{id}:
    get:
      tags: [Customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/jobs:
    get:
      tags: [Jobs]
      summary: List jobs
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/search'
        - name: customer_id
          in: query
          schema:
            type: integer
        - name: job_status
          in: query
          schema:
            type: string
        - name: business_unit_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
  /v1/jobs/{id}:
    get:
      tags: [Jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/audit:
    get:
      tags: [Audit]
      summary: Recent changes
      description: Returns recent audit log entries (last 7 days).
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - name: table
          in: query
          schema:
            type: string
          description: Filter by table (e.g., pricebook_services)
        - name: record_id
          in: query
          schema:
            type: string
          description: Filter by record ID
        - name: actor_type
          in: query
          schema:
            type: string
            enum: [user, sync_worker, outbound_worker, api, system]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditListResponse'
  /v1/audit/record/{table}/{id}:
    get:
      tags: [Audit]
      summary: Record history
      description: Returns complete change history for a specific record.
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
          description: Table name (e.g., pricebook_services)
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Record ID
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordHistoryResponse'
...
components:
  parameters:
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 100
        maximum: 1000
      description: Maximum records to return (max 1000)
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
      description: Number of records to skip
    search:
      name: search
      in: query
      schema:
        type: string
      description: Search term (searches name, code, description)
    active_only:
      name: active_only
      in: query
      schema:
        type: boolean
      description: Only return active records
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Resource not found"
    ListMeta:
      type: object
      required: [total, limit, offset, has_more]
      properties:
        total:
          type: integer
          example: 2332
        limit:
          type: integer
          example: 100
        offset:
          type: integer
          example: 0
        has_more:
          type: boolean
          example: true
        last_sync:
          type: string
          format: date-time
          example: "2026-01-21T18:30:00Z"
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        st_id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [Residential, Commercial]
        email:
          type: string
        phone:
          type: string
        active:
          type: boolean
        synced_at:
          type: string
          format: date-time
    CustomerListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
        meta:
          $ref: '#/components/schemas/ListMeta'
    CustomerResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Customer'
    Job:
      type: object
      properties:
        id:
          type: integer
        st_id:
          type: integer
        number:
          type: string
        customer_id:
          type: integer
        location_id:
          type: integer
        job_type_id:
          type: integer
        business_unit_id:
          type: integer
        job_status:
          type: string
        summary:
          type: string
        synced_at:
          type: string
          format: date-time
    JobListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Job'
        meta:
          $ref: '#/components/schemas/ListMeta'
    JobResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Job'
...
    AuditEntry:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        schema_name:
          type: string
        table_name:
          type: string
        operation:
          type: string
          enum: [INSERT, UPDATE, DELETE]
        record_id:
          type: string
        changed_fields:
          type: array
          items:
            type: string
        actor_type:
          type: string
        actor_id:
          type: string
        mutation_id:
          type: string
          format: uuid
        sync_batch_id:
          type: string
          format: uuid
    AuditListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditEntry'
    RecordHistoryResponse:
      type: object
      properties:
        record:
          type: object
          properties:
            schema:
              type: string
            table:
              type: string
            id:
              type: string
        data:
          type: array
          items:
            type: object
            properties:
              timestamp:
                type: string
                format: date-time
              operation:
                type: string
              changed_fields:
                type: array
                items:
                  type: string
              old_data:
                type: object
              new_data:
                type: object
              actor_type:
                type: string
              actor_id:
                type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for authentication
security:
  - ApiKeyAuth: []